sequence added and id turned to long

+++ src/main/java/com/id3/springboot/crudapi/dao/PersonDAO.java
@@ -8,9 +8,9 @@ public interface PersonDAO {
 
     public List<Person> findAll();
 
-    public Person findById(int id);
+    public Person findById(Long id);
 
     public void save(Person person);
 
-    public void deleteById(int id);
+    public void deleteById(Long id);
 }
